Mark directories as viewed when all their nested files/folders are viewed
