+++ src/js/components/file/index.jsx
@@ -62,7 +62,7 @@ class File extends React.Component {
         ) : null}
         {isViewed ? (
           <svg className='github-pr-file-checkmark octicon octicon-check' viewBox='0 0 12 16' version='1.1' width='12' height='16' aria-hidden='true'>
-            <path fill-rule='evenodd' d='M12 5l-8 8-4-4 1.5-1.5L4 10l6.5-6.5L12 5z' />
+            <path fillRule='evenodd' d='M12 5l-8 8-4-4 1.5-1.5L4 10l6.5-6.5L12 5z' />
           </svg>
         ) : null}
       </div>
